Rename useColor hooks for clarity (#14)
* Use colorScheme instead of corporateColor where unnecessary

* Rename useColorScheme for clarity

---------
