+++ src/app/(tabs)/action.tsx
@@ -1,13 +1,13 @@
 import { ScrollView } from "react-native-gesture-handler";
 
 import Colors from "#/constants/Colors";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 import AchievementComponent from "#/screens/ActionTab/components/AchievementComponent";
 import RegionMap from "#/screens/ActionTab/components/RegionMap";
 import StatisticsView from "#/screens/ActionTab/components/statistics/StatisticsView";
 
 const ActionTab = () => {
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   const backgroundColor = Colors[colorScheme].secondaryBackground;
   return (
     <ScrollView style={{ flex: 1, backgroundColor }}>
