+++ src/app/licenses.tsx
@@ -3,11 +3,11 @@ import { View } from "react-native";
 import NavBar from "#/components/bars/NavBar";
 import Colors from "#/constants/Colors";
 import { styles } from "#/constants/Styles";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 import Licenses from "#/screens/Settings/components/licenses/Licenses";
 
 const LicensesScreen = () => {
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   const backgroundColor = Colors[colorScheme].background;
   return (
     <View
