+++ src/app/search.tsx
@@ -6,7 +6,7 @@ import { StyleProp, TextInput, TextStyle } from "react-native";
 import NavBar from "#/components/bars/NavBar";
 import View from "#/components/design/View";
 import Colors from "#/constants/Colors";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 import AISearch from "#/screens/Search/components/AISearch";
 import AlgoliaSearchResults from "#/screens/Search/components/AlgoliaSearch";
 import SearchHeader from "#/screens/Search/components/SearchHeader";
@@ -18,7 +18,7 @@ const SearchScreen = () => {
   const parameters = useLocalSearchParams<{ tag: string }>();
   const tag: string | undefined = parameters?.tag ?? shareIntent?.webUrl;
   const searchReference = useRef<TextInput>(null);
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   const backgroundColor = Colors[colorScheme].background;
 
   // Focus the search input when the screen is focused (unless there's a tag)
