+++ src/app/support.tsx
@@ -6,11 +6,11 @@ import Support from "#/components/views/Support";
 import Colors from "#/constants/Colors";
 import Config from "#/constants/Config";
 import { styles as globalStyles } from "#/constants/Styles";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 
 const SupportScreen = () => {
   const wpUrl = Config.wpUrl;
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   const backgroundColor = Colors[colorScheme].secondaryBackground;
   return (
     <View style={globalStyles.container}>
