+++ src/components/animations/AnimatedLoading.tsx
@@ -3,7 +3,7 @@ import { ActivityIndicator, Image } from "react-native";
 import View from "#/components/design/View";
 import { styles } from "#/constants/Styles";
 import { isVolksverpetzer } from "#/helpers/utils/variant";
-import { useCorporateColor } from "#/hooks/useColorScheme";
+import { useCorporateColor } from "#/hooks/useAppColorScheme";
 
 import LogoAnimation from "#assets/images/LogoAnimation4_1.gif";
