+++ src/components/popups/MissionPopup.tsx
@@ -7,7 +7,7 @@ import Text from "#/components/design/Text";
 import View from "#/components/design/View";
 import Colors from "#/constants/Colors";
 import { styles } from "#/constants/Styles";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 
 interface MissionPopupProperties {
   text1: string;
@@ -15,7 +15,7 @@ interface MissionPopupProperties {
 }
 
 const MissionPopup = ({ text1, text2 }: MissionPopupProperties) => {
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   const backgroundColor = Colors[colorScheme].background;
   const corporate = Colors[colorScheme].corporate;
   const router = useRouter();
