+++ src/components/posts/ClaimPost.tsx
@@ -4,7 +4,7 @@ import { TouchableOpacity } from "react-native";
 import Text from "#/components/design/Text";
 import View from "#/components/design/View";
 import Colors from "#/constants/Colors";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 
 interface Review {
   publisher: {
@@ -34,7 +34,7 @@ export interface ClaimProperties {
  * Renders Short Post for Googles Claim Review
  */
 const ClaimPost = (properties: ClaimProperties) => {
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   const color = Colors[colorScheme].grayedOutText;
   const handleSelectClaim = () => {
     WebBrowser.openBrowserAsync(properties.claimReview[0].url);
