+++ src/components/typography/Heading.tsx
@@ -3,7 +3,7 @@ import { Text, TextProps, TextStyle } from "react-native";
 
 import Colors from "#/constants/Colors";
 import { styles as _styles } from "#/constants/Styles";
-import useColorScheme from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 
 interface HeadingProperties extends TextProps {
   children: ReactNode;
@@ -12,7 +12,7 @@ interface HeadingProperties extends TextProps {
 
 const Heading = (properties: HeadingProperties) => {
   const { styles, children } = properties;
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   return (
     <Text
       style={{
