+++ src/hooks/useAppColorScheme.ts
@@ -1,8 +1,5 @@
 import { useContext } from "react";
-import {
-  ColorSchemeName,
-  useColorScheme as _useColorScheme,
-} from "react-native";
+import { ColorSchemeName, useColorScheme } from "react-native";
 
 import Colors from "#/constants/Colors";
 import SettingsContext from "#/helpers/SettingsContext";
@@ -11,9 +8,9 @@ import SettingsContext from "#/helpers/SettingsContext";
  * A hook that returns the color scheme of the device, or "dark" if the user has set the advanced setting "Always Dark Mode".
  * @returns The color scheme of the device, or "dark" if the user has set the advanced setting "Always Dark Mode".
  */
-export default function useColorScheme(): NonNullable<ColorSchemeName> {
+export default function useAppColorScheme(): NonNullable<ColorSchemeName> {
   const { advancedSettings } = useContext(SettingsContext);
-  const scheme = _useColorScheme();
+  const scheme = useColorScheme();
   if (advancedSettings?.alwaysDarkMode?.value ?? false)
     return "dark" as NonNullable<ColorSchemeName>;
   return scheme as NonNullable<ColorSchemeName>;
@@ -24,11 +21,11 @@ export default function useColorScheme(): NonNullable<ColorSchemeName> {
  * @returns The corporate color for the current color scheme.
  */
 export const useCorporateColor = () => {
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   return Colors[colorScheme].corporate;
 };
 
 export const useReverseColorScheme = () => {
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
   return colorScheme === "dark" ? "light" : "dark";
 };
