+++ src/hooks/useThemeColor.tsx
@@ -1,12 +1,12 @@
-import Colors from "#/constants/Colors";
+import useAppColorScheme from "src/hooks/useAppColorScheme";
 
-import useColorScheme from "./useColorScheme";
+import Colors from "#/constants/Colors";
 
 export function useThemeColor(
   properties: { light?: string; dark?: string },
   colorName: keyof typeof Colors.light & keyof typeof Colors.dark,
 ) {
-  const theme = useColorScheme();
+  const theme = useAppColorScheme();
   const colorFromProperties = properties[theme];
 
   return colorFromProperties || Colors[theme][colorName];
