+++ src/screens/Onboarding/components/Stepper.tsx
@@ -8,7 +8,7 @@ import {
 import Text from "#/components/design/Text";
 import View from "#/components/design/View";
 import { styles } from "#/constants/Styles";
-import { useCorporateColor } from "#/hooks/useColorScheme";
+import { useCorporateColor } from "#/hooks/useAppColorScheme";
 
 import { OnBoardingData } from "./Flatboard";
