+++ src/screens/Search/components/AlgoliaSearch.tsx
@@ -16,7 +16,7 @@ import Text from "#/components/design/Text";
 import View from "#/components/design/View";
 import Colors from "#/constants/Colors";
 import { onLinkPress } from "#/helpers/Linking";
-import useColorScheme, { useCorporateColor } from "#/hooks/useColorScheme";
+import useAppColorScheme from "#/hooks/useAppColorScheme";
 
 interface AlgoliaSearchProperties {
   searchString: string;
@@ -34,8 +34,8 @@ const AlgoliaSearchResults = ({
 
   // Initialize Algolia client
   const client = searchClient("W8YO8C6SIN", "f8211e7620b2d30da0d73f451fe36634");
-  const highlightColor = useCorporateColor();
-  const colorScheme = useColorScheme();
+  const colorScheme = useAppColorScheme();
+  const highlightColor = Colors[colorScheme].corporate;
   const textColor = Colors[colorScheme].text;
 
   // Create a debounced search function to avoid too many API calls
